requirements.txt
requests==2.32.3
pandas==2.2.2
pyarrow==17.0.0
python-dotenv==1.0.1
web3==7.1.0